+++ API/app.py
@@ -1,0 +1,9 @@
+from flask import Flask
+#from calc... import auth_bp
+
+app = Flask(__name__)
+
+#app.register_blueprint(auth_bp, url_prefix='/auth')
+
+if __name__ == '__main__':
+    app.run(debug=True)
